Extract CustomerMessageForm feature from customerList
