+++ src/features/customerMessageForm/types.ts
@@ -1,0 +1,10 @@
+export interface MessageTemplate {
+  id: string;
+  name: string;
+}
+
+export interface MessageFormValues {
+  templateId: string;
+  subject: string;
+  message: string;
+}
